requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pydantic==2.1.1
python-dotenv==1.0.1
requests==2.32.3
nltk==3.9.1
transformers==4.53.2
torch==2.5.1
xgboost==3.0.0
scikit-learn==1.5.1
numpy==1.26.4
pandas==2.2.3
openai==1.3.7
